move plot-related CLI to plot submod
